// src/shared/api/generated/schemas/adminUsersListResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Smetchik API
 * Smetchik service API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { AdminUser } from './adminUser';

export interface AdminUsersListResponse {
  users?: AdminUser[];
  total?: number;
  limit?: number;
  offset?: number;
}